+++ src/Plane/Curves/Ellipse.php
@@ -10,19 +10,50 @@ use samizdam\Geometry\Plane\Lines\LineSegmentInterface;
  * @author samizdam
  *        
  */
-class Ellipse extends Circle implements EllipseInterface, FactoriesCollectionAwareInterface
+class Ellipse implements EllipseInterface, FactoriesCollectionAwareInterface
 {
     
     use FactoriesCollectionAwareTrait;
 
+    private $innerCircle;
+
+    private $minorAxis;
+
+    private $majorAxis;
+
     public function __construct(LineSegmentInterface $minorAxis, LineSegmentInterface $majorAxis)
-    {}
+    {
+        $this->minorAxis = $minorAxis;
+        $this->majorAxis = $majorAxis;
+        
+        $centralPoint = $minorAxis->getCentralPoint();
+        $this->innerCircle = new Circle($centralPoint, $minorAxis->getLength() / 2);
+    }
+
+    public function getCentralPoint()
+    {
+        return $this->innerCircle->getCentralPoint();
+    }
 
     public function getMajorAxis()
     {
-//         $this->getFactoriesCollection()->getLineFactory()->
+        return $this->majorAxis;
+        // $this->getFactoriesCollection()->getLineFactory()->
     }
 
     public function getMinorAxis()
+    {
+        return $this->minorAxis;
+    }
+
+    public function getLength()
     {}
+
+    public function getArea()
+    {}
+
+    public function getR()
+    {
+        return $this->minorAxis->getLength();
+    }
 }