+++ src/Plane/Lines/LineSegment.php
@@ -2,11 +2,13 @@
 namespace samizdam\Geometry\Plane\Lines;
 
 use samizdam\Geometry\Plane\CalculatorAwareTrait;
+use samizdam\Geometry\Plane\FactoriesCollectionAwareTrait;
 
 class LineSegment extends AbstractLine implements LineSegmentInterface
 {
     
     use CalculatorAwareTrait;
+    use FactoriesCollectionAwareTrait;
 
     public function getFirstPoint()
     {
@@ -20,6 +22,20 @@ class LineSegment extends AbstractLine implements LineSegmentInterface
 
     public function getLength()
     {
-        return $this->getComposeCalculator()->getCalculator(LengthCalculatorInterface::class)->getSegmentLength($this);
+        return $this->getComposeCalculator()
+            ->getCalculator(LengthCalculatorInterface::class)
+            ->getSegmentLength($this);
+    }
+
+    public function getCentralPoint()
+    {
+        $Ax = $this->getFirstPoint()->getX();
+        $Ay = $this->getFirstPoint()->getY();
+        $Bx = $this->getLastPoint()->getX();
+        $By = $this->getLastPoint()->getY();
+        
+        $Cx = ($Ax + $Bx) / 2;
+        $Cy = ($Ay + $By) / 2;
+        return $this->getFactoriesCollection()->getPoint($Cx, $Cy);
     }
 }